test: cover connecting prompt shortcut
