+++ packages/web/src/app/(app)/session/[id]/page.tsx
@@ -28,14 +28,9 @@ import {
   type SessionListResponse,
 } from "@/lib/session-list";
 import { useMediaQuery } from "@/hooks/use-media-query";
-import type { SessionAttachmentReference } from "@open-inspect/shared/types/session-attachments";
 import { DEFAULT_MODEL, getDefaultReasoningEffort } from "@open-inspect/shared/models";
 import { resolveModelPreference, type ModelPreference } from "@/lib/model-selection";
 import { useEnabledModels } from "@/hooks/use-enabled-models";
-import {
-  DEFAULT_ATTACHMENT_ONLY_MESSAGE,
-  useSessionAttachments,
-} from "@/hooks/use-session-attachments";
 import type { ComboboxGroup } from "@/components/ui/combobox";
 import { useSessionDiffs } from "@/hooks/use-session-diffs";
 import { resolveDiffSelection, type DiffSelection } from "@/lib/session-diffs";
@@ -54,22 +49,18 @@ import { focusSessionDetailsTrigger } from "@/lib/session-details-focus";
 import { useSessionParticipantProfiles } from "@/hooks/use-session-participant-profiles";
 import { useSessionDetailsSidebar } from "@/hooks/use-session-details-sidebar";
 import {
-  promptRequestSignature,
-  resolvePromptRequestIdentity,
-  type PromptRequestIdentity,
-} from "@/lib/prompt-request-id";
-import {
   classifySessionReadAttempt,
   markMessageRead,
   reconcileSessionReadState,
   SessionReadRequestError,
 } from "@/lib/session-read-state";
-import { restoreQueuedPrompt } from "@/lib/restore-queued-prompt";
+import { usePromptInput } from "@/hooks/use-prompt-input";
 import { useSessionSnapshot } from "./session-snapshot-provider";
 
 type SessionState = ReturnType<typeof useSessionSocket>["sessionState"];
 
 const TERMINAL_VISIBLE_STORAGE_KEY = "terminal-visible";
+const DEFAULT_SESSION_STATUS = "created" as const;
 
 export default function SessionPage() {
   const initialSnapshot = useSessionSnapshot();
@@ -135,7 +126,7 @@ export default function SessionPage() {
     selectedModel,
     reasoningEffort,
     loadingEnabledModels,
-    sessionState?.status ?? "created",
+    sessionState?.status ?? DEFAULT_SESSION_STATUS,
     ready
   );
   const [cancellingPromptIds, setCancellingPromptIds] = useState<ReadonlySet<string>>(new Set());
@@ -351,7 +342,7 @@ export default function SessionPage() {
       <SessionPromptComposer
         session={{
           id: sessionId,
-          status: sessionState?.status ?? "created",
+          status: sessionState?.status ?? DEFAULT_SESSION_STATUS,
           artifacts,
           primaryRepo,
           onArchive: handleArchive,
@@ -404,7 +395,7 @@ export default function SessionPage() {
         onOpenMobileDetails={openMobileDetails}
         actions={{
           sessionId,
-          sessionStatus: sessionState?.status ?? "created",
+          sessionStatus: sessionState?.status ?? DEFAULT_SESSION_STATUS,
           artifacts,
           primaryRepo,
           onArchive: handleArchive,
@@ -688,176 +679,5 @@ function useModelSelection(sessionState: SessionState) {
     handleModelChange,
     modelItems,
     loadingEnabledModels,
-  };
-}
-
-/**
- * Prompt textarea state and handlers: submit, Cmd/Ctrl+Enter, and the
- * debounced typing indicator.
- */
-function usePromptInput(
-  sessionId: string,
-  sendPrompt: ReturnType<typeof useSessionSocket>["sendPrompt"],
-  sendTyping: ReturnType<typeof useSessionSocket>["sendTyping"],
-  selectedModel: string,
-  reasoningEffort: string | undefined,
-  loadingEnabledModels: boolean,
-  sessionStatus: NonNullable<SessionState>["status"],
-  canSubmit: boolean
-) {
-  const [prompt, setPromptState] = useState("");
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [submitError, setSubmitError] = useState<string | null>(null);
-  const sessionAttachments = useSessionAttachments();
-  const hasAttachments = sessionAttachments.hasAttachments;
-  const inputRef = useRef<HTMLTextAreaElement>(null);
-  const typingTimeoutRef = useRef<NodeJS.Timeout | null>(null);
-  const submitInFlightRef = useRef(false);
-  const restoreFocusAfterSubmitRef = useRef(false);
-  const retryRequestRef = useRef<PromptRequestIdentity | null>(null);
-  const attachmentDraftSignature = sessionAttachments.attachments
-    .map((attachment) => attachment.id)
-    .join("\u0000");
-  const promptRef = useRef(prompt);
-  const setPrompt = useCallback((value: string) => {
-    promptRef.current = value;
-    setPromptState(value);
-  }, []);
-
-  const clearTypingTimeout = useCallback(() => {
-    if (typingTimeoutRef.current) {
-      clearTimeout(typingTimeoutRef.current);
-      typingTimeoutRef.current = null;
-    }
-  }, []);
-
-  useEffect(() => clearTypingTimeout, [clearTypingTimeout]);
-  useEffect(() => {
-    retryRequestRef.current = null;
-  }, [selectedModel, reasoningEffort, attachmentDraftSignature]);
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    const hasAttachments = sessionAttachments.attachments.length > 0;
-    if (
-      submitInFlightRef.current ||
-      !canSubmit ||
-      (!prompt.trim() && !hasAttachments) ||
-      sessionStatus === "archived" ||
-      sessionStatus === "cancelled" ||
-      loadingEnabledModels ||
-      sessionAttachments.isUploading
-    ) {
-      return;
-    }
-
-    submitInFlightRef.current = true;
-    restoreFocusAfterSubmitRef.current = document.activeElement === inputRef.current;
-    setIsSubmitting(true);
-    setSubmitError(null);
-    try {
-      const content = prompt.trim() || DEFAULT_ATTACHMENT_ONLY_MESSAGE;
-      let attachments: SessionAttachmentReference[] | undefined;
-      if (hasAttachments) {
-        try {
-          attachments = await sessionAttachments.uploadAll(sessionId);
-        } catch (error) {
-          setSubmitError(error instanceof Error ? error.message : "Failed to upload attachments");
-          return;
-        }
-      }
-
-      // Drop any queued typing indicator — the prompt supersedes it
-      clearTypingTimeout();
-      const signature = promptRequestSignature({
-        content,
-        model: selectedModel,
-        reasoningEffort,
-        attachmentIds: sessionAttachments.attachments.map((attachment) => attachment.id),
-      });
-      const requestIdentity = resolvePromptRequestIdentity(signature, retryRequestRef.current);
-      retryRequestRef.current = requestIdentity;
-      const result = await sendPrompt(
-        content,
-        selectedModel,
-        reasoningEffort,
-        attachments,
-        requestIdentity.clientRequestId
-      );
-      if (!result.ok) {
-        setSubmitError(
-          result.message ??
-            (result.reason === "timeout"
-              ? "Confirmation timed out. Retry while this page is open to reuse the same request."
-              : result.reason === "disconnected"
-                ? "Disconnected before confirmation. Retry on this page after reconnecting."
-                : "The prompt could not be queued.")
-        );
-        return;
-      }
-
-      retryRequestRef.current = null;
-      setPrompt("");
-      sessionAttachments.clearAttachments();
-      // Revalidate sidebar so this session bubbles to the top
-      mutate(isUnarchivedSessionListKey);
-    } finally {
-      submitInFlightRef.current = false;
-      setIsSubmitting(false);
-      if (restoreFocusAfterSubmitRef.current) {
-        restoreFocusAfterSubmitRef.current = false;
-        requestAnimationFrame(() => {
-          if (document.activeElement === document.body) inputRef.current?.focus();
-        });
-      }
-    }
-  };
-
-  const handleKeyDown = (e: React.KeyboardEvent) => {
-    if (e.nativeEvent.isComposing) return;
-
-    if (e.key === "Enter" && (e.metaKey || e.ctrlKey) && !e.shiftKey && !e.altKey) {
-      e.preventDefault();
-      handleSubmit(e);
-    }
-  };
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    setPrompt(e.target.value);
-    setSubmitError(null);
-    retryRequestRef.current = null;
-
-    // Send typing indicator (debounced)
-    clearTypingTimeout();
-    typingTimeoutRef.current = setTimeout(() => {
-      sendTyping();
-    }, 300);
-  };
-
-  const restorePrompt = useCallback(
-    (content: string) => {
-      const restored = restoreQueuedPrompt({
-        content,
-        currentPrompt: promptRef.current,
-        hasAttachments: hasAttachments(),
-        setPrompt,
-        input: inputRef.current,
-      });
-      return restored;
-    },
-    [hasAttachments, setPrompt]
-  );
-
-  return {
-    prompt,
-    sessionAttachments,
-    inputRef,
-    isSubmitting,
-    submitError,
-    setSubmitError,
-    handleSubmit,
-    handleInputChange,
-    handleKeyDown,
-    restorePrompt,
   };
 }
